+++ src/renderer/CategoryPage.tsx
@@ -509,6 +509,7 @@ const CategoryPage = (props: IProps) => {
                 setAppState={setAppState}
                 table={table}
                 stateManager={stateManager}
+                filteredState={filteredState}
               />
             </div>
           </div>